test(provider/memory): added more test

+++ pkg/hyperpic/provider/memory/configuration.go
@@ -10,10 +10,10 @@ import "time"
 type CacheConfiguration struct {
 	LifeTime      time.Duration `mapstructure:"life_time"`
 	CleanInterval time.Duration `mapstructure:"clean_interval"`
-	MemoryLimit   uint64        `mapstructure:"memory_limit"`
+	MemoryLimit   int           `mapstructure:"memory_limit"`
 }
 
 // SourceConfiguration struct
 type SourceConfiguration struct {
-	MemoryLimit uint64 `mapstructure:"memory_limit"`
+	MemoryLimit int `mapstructure:"memory_limit"`
 }
